+++ devices/device-even/src/iosMain/kotlin/com/xgglass/device/even/EvenIosGlassesClient.kt
@@ -565,11 +565,11 @@ class EvenIosGlassesClient(
 
         when (val event = EvenStateEvents.stateEvent(value)) {
             is EvenStateEvent.Tap -> {
-                _events.tryEmit(GlassesEvent.Tap(event.count))
+                emitEvent(GlassesEvent.Tap(event.count))
                 return
             }
             EvenStateEvent.LongPress -> {
-                _events.tryEmit(GlassesEvent.LongPress)
+                emitEvent(GlassesEvent.LongPress)
                 return
             }
             EvenStateEvent.AiRecordOver -> {